@@ -5,7 +5,7 @@ author: fiissh.zhao <zhaoyongchun2010@gmail.com>
 homepage: https://github.com/fiissh/flutter_file_path_provider
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
